Module Catapult

Catapult-based tracing.

A nice profiling format based on json, useful for visualizing what goes on. This library is the instrumentation part; see catapult-client, catapult-file, or use a custom BACKEND to actually record traces.

If a trace is obtained in, say, the file "trace.json.gz", it can be opened in chrome/chromium at "chrome://tracing".

Tracy can import (uncompressed) trace files with a nice native trace explorer.

See the documentation of TEF

module type BACKEND = sig ... end

module type IMPL = sig ... end

module Arg : sig ... end

Custom arguments.

module Control = Tracing.Control

module Event_type : sig ... end

Event types.

module Nil_impl : sig ... end

Nil implementation.

module Tracing : sig ... end

Profiling probes.