Module type Catapult.BACKEND

val emit : id:string option -> name:string -> ph:Event_type.t -> tid:int -> pid:int -> cat:string list option -> ts_us:float -> args:(string * Arg.t) list option -> stack:string list option -> dur:float option -> ?extra:(string * string) list -> unit -> unit

Emit an event.

val tick : unit -> unit

Function that can be called regularly to ensure background work is done. Potentially useful for long-running processes.

val teardown : unit -> unit

Tear down backend once the program is shutting down.